MLT-31: Limit all endpoints and xwork to the confluence Admin only

+++ src/main/java/com/ntnguyen/app/confluence/licensetracking/rest/v1/SubscriberController.java
@@ -3,6 +3,9 @@ package com.ntnguyen.app.confluence.licensetracking.rest.v1;
 import static com.ntnguyen.app.confluence.licensetracking.mapper.SubscriberMapper.subscriberEntityToDto;
 
 import com.atlassian.annotations.security.XsrfProtectionExcluded;
+import com.atlassian.confluence.security.PermissionManager;
+import com.atlassian.confluence.user.AuthenticatedUserThreadLocal;
+import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.ntnguyen.app.confluence.licensetracking.persistent.entity.SubscriberEntity;
 import com.ntnguyen.app.confluence.licensetracking.service.ConfigurationService;
@@ -24,11 +27,15 @@ import org.springframework.stereotype.Component;
 @Path("/subscribers")
 public class SubscriberController {
 
-  private static final String EMAIL_PATTERN = "(([^<>()\\[\\]\\\\.,;:\\s@\"]+(\\.[^<>()\\[\\]\\\\.,;:\\s@\"]+)*)|(\".+\"))@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\])|(([a-zA-Z\\-0-9]+\\.)+[a-zA-Z]{2,}))";
+  private static final String EMAIL_PATTERN = "(([^<>()\\[\\]\\\\.,;:\\s@\"]+(\\.[^<>()\\[\\]\\\\.,;:\\s@\"]+)*)|(\".+\"))@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}])|(([a-zA-Z\\-0-9]+\\.)+[a-zA-Z]{2,}))";
+
+  private final PermissionManager permissionManager;
   private final ConfigurationService configurationService;
 
   @Autowired
-  public SubscriberController(ConfigurationService configurationService) {
+  public SubscriberController(@ComponentImport PermissionManager permissionManager,
+      ConfigurationService configurationService) {
+    this.permissionManager = permissionManager;
     this.configurationService = configurationService;
   }
 
@@ -38,6 +45,12 @@ public class SubscriberController {
   @Produces({MediaType.APPLICATION_JSON})
   @XsrfProtectionExcluded
   public Response addSubscriber(@FormParam("email") String email) throws JsonProcessingException {
+    if (isNotAdmin()) {
+      return Response.status(Status.FORBIDDEN)
+          .entity("{\"message\": \"You are not allow to perform this operation\"}")
+          .build();
+    }
+
     if (!validEmail(email)) {
       return Response.status(Status.BAD_REQUEST)
           .entity("{\"message\": \"Email is not valid\"}")
@@ -62,6 +75,12 @@ public class SubscriberController {
   @Produces({MediaType.APPLICATION_JSON})
   @XsrfProtectionExcluded
   public Response deleteSubscriber(@PathParam("email") String email) {
+    if (isNotAdmin()) {
+      return Response.status(Status.FORBIDDEN)
+          .entity("{\"message\": \"You are not allow to perform this operation\"}")
+          .build();
+    }
+
     if (isSubscriberAlreadySaved(email)) {
       configurationService.deleteSubscriber(email);
       return Response.ok("{"
@@ -86,5 +105,9 @@ public class SubscriberController {
 
   private boolean isSubscriberAlreadySaved(String email) {
     return configurationService.getSubscriber(email) != null;
+  }
+
+  private boolean isNotAdmin() {
+    return !permissionManager.isConfluenceAdministrator(AuthenticatedUserThreadLocal.get());
   }
 }
